backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
prophet==1.1.5
pandas==2.1.3
python-dotenv==1.0.0
pydantic==2.5.0
gunicorn==20.1.0